# Conflicts:
#	binkapp.xcodeproj/project.pbxproj
#	binkapp/BinkNetworking/ApiManager.swift
#	binkapp/BinkUI/BinkGradientButton.swift
#	binkapp/Screens/AddingOptions/AddingOptionsViewController.swift
#	binkapp/Screens/AddingOptions/AddingOptionsViewModel.swift

+++ binkapp/Screens/LoyaltyWallet/Models/MembershipTransaction.swift
@@ -10,7 +10,7 @@ import Foundation
 struct MembershipTransaction: Codable {
     let id: Int?
     let status: String?
-    let timestamp: Int?
+    let timestamp: Double?
     let description : String?
     let amounts: [MembershipCardAmount]?
     
@@ -27,7 +27,7 @@ struct MembershipTransaction: Codable {
         let values = try decoder.container(keyedBy: CodingKeys.self)
         id = try values.decodeIfPresent(Int.self, forKey: .id)
         status = try values.decodeIfPresent(String.self, forKey: .status)
-        timestamp = try values.decodeIfPresent(Int.self, forKey: .timestamp)
+        timestamp = try values.decodeIfPresent(Double.self, forKey: .timestamp)
         description = try values.decodeIfPresent(String.self, forKey: .description)
         amounts = try values.decodeIfPresent([MembershipCardAmount].self, forKey: .amounts)
     }
